Implement DataGenerator class
DataGenerator class is the new implementation used to generate
production-grade data sets. The .generate method, when using
a CHUNK_SIZE of 20,000 and MAX_ROWS count of 10,000,000, can generate
two .csv files of 10,000,000 lines each in roughly 49 seconds whilst
only consuming roughly 60mb of RAM in the process. This was achieved
by implementing both a Readable and Writable stream to reduce the load
on memory at any given point in time.

This is a huge improvement from before as the solution using only
Writable streams took roughly 90 seconds to run and consumed roughly
1.5GB of RAM in the process.